+++ shuup/core/utils/price_cache.py
@@ -7,7 +7,9 @@
 """
 Utilities for caching price info
 """
-from shuup.core.models import ShopProduct
+from decimal import Decimal
+
+from shuup.core.models import AnonymousContact, ShopProduct
 from shuup.core.pricing import PriceInfo
 from shuup.core.utils import context_cache
 from shuup.utils.dates import to_timestamp
@@ -21,13 +23,39 @@ def _get_price_info_namespace_for_shop(shop_id):
 
 def _get_price_info_cache_key_params(context, item, quantity, **context_args):
     shop_id = context.shop.pk if hasattr(context, "shop") else 0
+    customer = getattr(context, "customer", None)
+
+    if customer:
+        cached_customer_groups = getattr(customer, "_cached_customer_groups", None)
+        if cached_customer_groups:
+            groups = customer._cached_customer_groups
+        else:
+            groups = list(customer.groups.order_by("pk").values_list("pk", flat=True))
+            customer._cached_customer_groups = groups
+    else:
+        anonymous_group_id = getattr(AnonymousContact, "_cached_default_group_id", None)
+        if anonymous_group_id:
+            groups = [AnonymousContact._cached_default_group_id]
+        else:
+            anonymous_group_id = AnonymousContact().default_group.pk
+            AnonymousContact._cached_default_group_id = anonymous_group_id
+            groups = [anonymous_group_id]
+
+    extra_kwargs = dict()
+    for key, value in context_args.items():
+        if hasattr(value, "pk"):
+            extra_kwargs[key] = value.pk
+        else:
+            extra_kwargs[key] = value
+
     return dict(
         identifier="price_info_cache",
         item=_get_price_info_namespace_for_shop(shop_id),
-        context={"customer": getattr(context, "customer", None)},
-        quantity=quantity,
-        context_item=item,
-        **context_args
+        context={},
+        customer_groups=groups,
+        quantity=str(Decimal(quantity)),
+        item_id=item.pk if hasattr(item, "pk") else str(item),
+        **extra_kwargs
     )
 
 
@@ -50,9 +78,8 @@ def cache_many_price_info(context, item, quantity, prices_infos, **context_args)
     if not all(isinstance(item, PriceInfo) for item in prices_infos):
         return
 
-    key = context_cache.get_cache_key_for_context(
-        many=True, **_get_price_info_cache_key_params(context, item, quantity, **context_args)
-    )
+    context_kwargs = _get_price_info_cache_key_params(context, item, quantity, **context_args)
+    key = context_cache.get_cache_key_for_context(many=True, **context_kwargs)
     context_cache.set_cached_value(key, prices_infos)
 
 
@@ -81,9 +108,8 @@ def get_many_cached_price_info(context, item, quantity=1, **context_args):
     :param object item
     :param float|Decimal quantity
     """
-    key, prices_infos = context_cache.get_cached_value(
-        many=True, **_get_price_info_cache_key_params(context, item, quantity, **context_args)
-    )
+    cache_kwargs = _get_price_info_cache_key_params(context, item, quantity, many=True, **context_args)
+    key, prices_infos = context_cache.get_cached_value(**cache_kwargs)
 
     if prices_infos:
         try:
@@ -112,9 +138,8 @@ def get_cached_price_info(context, item, quantity=1, **context_args):
     :param object item
     :param float|Decimal quantity
     """
-    key, price_info = context_cache.get_cached_value(
-        **_get_price_info_cache_key_params(context, item, quantity, **context_args)
-    )
+    cache_kwargs = _get_price_info_cache_key_params(context, item, quantity, **context_args)
+    key, price_info = context_cache.get_cached_value(**cache_kwargs)
 
     from django.utils.timezone import now