+++ CHANGELOG.md
@@ -8,8 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+### Fixed
+
+- General: replace hash() with sha1() as Python's hash() function doesn't have a stable result across processes
+- Core: use SHA-1 to hash cache keys as Python's hash() function doesn't have a stable result across processes
+
 ### Changed
 
+- Cache product orderability and prices using a list of user groups instead of per contact
 - Pull translation strings from Transifex
 
 ## [2.9.1] - 2021-05-24